fix: rename dialog components
