+++ src/components/dialog/DialogEncrypt.js
