+++ src/components/dialog/DialogManualAlert.js
