+++ src/pages/Appearance/Datasets/SpecificTableAction.js
@@ -3,7 +3,7 @@ import i18n from '@dhis2/d2-i18n'
 import PropTypes from '@dhis2/prop-types'
 import { datasetHasCategoryCombo } from './helper'
 import TableActions from '../../../components/table-actions'
-import DialogDelete from '../../../components/dialog/dialog-delete'
+import DialogDelete from '../../../components/dialog/DialogDelete'
 import DialogDatasetSpecificSetting from './DialogDatasetSpecificSetting'
 import {
     removeSettingsFromList,
