+++ src/pages/Appearance/Programs/SpecificTableAction.js
@@ -2,7 +2,7 @@ import React, { useState } from 'react'
 import i18n from '@dhis2/d2-i18n'
 import PropTypes from '@dhis2/prop-types'
 import TableActions from '../../../components/table-actions'
-import DialogDelete from '../../../components/dialog/dialog-delete'
+import DialogDelete from '../../../components/dialog/DialogDelete'
 import DialogNewProgram from './DialogNewProgram'
 import { programHasCategoryCombo } from './helper'
 import {
